# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "AdminLoginChallenge": {
        "properties": {
          "challenge_token": {
            "title": "Challenge Token",
            "type": "string"
          },
          "enrollment_required": {
            "title": "Enrollment Required",
            "type": "boolean"
          },
          "methods": {
            "items": {
              "enum": [
                "totp",
                "webauthn"
              ],
              "type": "string"
            },
            "title": "Methods",
            "type": "array"
          },
          "mfa_required": {
            "const": true,
            "default": true,
            "title": "Mfa Required",
            "type": "boolean"
          },
          "webauthn_challenge": {
            "title": "Webauthn Challenge",
            "type": "string"
          },
          "webauthn_credential_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Webauthn Credential Id"
          },
          "webauthn_rp_id": {
            "title": "Webauthn Rp Id",
            "type": "string"
          }
        },
        "required": [
          "challenge_token",
          "enrollment_required",
          "methods",
          "webauthn_challenge",
          "webauthn_rp_id"
        ],
        "title": "AdminLoginChallenge",
        "type": "object"
      },
      "AdminLoginRequest": {
        "additionalProperties": false,
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "maxLength": 128,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "AdminLoginRequest",
        "type": "object"
      },
      "AdminPosition": {
        "description": "The data-reach position carried by the single Admin role.",
        "enum": [
          "chairperson",
          "dean",
          "registrar"
        ],
        "title": "AdminPosition",
        "type": "string"
      },
      "AdminSessionRead": {
        "properties": {
          "college_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "College Id"
          },
          "department_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Department Id"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "position": {
            "$ref": "#/components/schemas/AdminPosition"
          },
          "role": {
            "const": "admin",
            "default": "admin",
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "position"
        ],
        "title": "AdminSessionRead",
        "type": "object"
      },
      "AdminTokenPair": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "college_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "College Id"
          },
          "department_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Department Id"
          },
          "position": {
            "$ref": "#/components/schemas/AdminPosition"
          },
          "refresh_token": {
            "title": "Refresh Token",
            "type": "string"
          },
          "role": {
            "const": "admin",
            "default": "admin",
            "title": "Role",
            "type": "string"
          },
          "token_type": {
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "refresh_token",
          "position"
        ],
        "title": "AdminTokenPair",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LoginRequest": {
        "additionalProperties": false,
        "description": "Payload for exchanging credentials for tokens.",
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "maxLength": 128,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
      "MfaVerifyRequest": {
        "additionalProperties": false,
        "properties": {
          "assertion": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assertion"
          },
          "challenge_token": {
            "title": "Challenge Token",
            "type": "string"
          },
          "code": {
            "anyOf": [
              {
                "maxLength": 6,
                "minLength": 6,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          }
        },
        "required": [
          "challenge_token"
        ],
        "title": "MfaVerifyRequest",
        "type": "object"
      },
...
      "RefreshRequest": {
        "additionalProperties": false,
        "description": "Payload for exchanging a refresh token for a new token pair.",
        "properties": {
          "refresh_token": {
            "title": "Refresh Token",
            "type": "string"
          }
        },
        "required": [
          "refresh_token"
        ],
        "title": "RefreshRequest",
        "type": "object"
      },
...
      "TokenPair": {
        "description": "An access/refresh token pair, as returned by login and refresh.",
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "refresh_token": {
            "title": "Refresh Token",
            "type": "string"
          },
          "token_type": {
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "refresh_token"
        ],
        "title": "TokenPair",
        "type": "object"
      },
      "TotpEnrollmentConfirmRequest": {
        "additionalProperties": false,
        "properties": {
          "challenge_token": {
            "title": "Challenge Token",
            "type": "string"
          },
          "code": {
            "maxLength": 6,
            "minLength": 6,
            "title": "Code",
            "type": "string"
          }
        },
        "required": [
          "challenge_token",
          "code"
        ],
        "title": "TotpEnrollmentConfirmRequest",
        "type": "object"
      },
      "TotpEnrollmentRead": {
        "properties": {
          "provisioning_uri": {
            "title": "Provisioning Uri",
            "type": "string"
          },
          "secret": {
            "title": "Secret",
            "type": "string"
          }
        },
        "required": [
          "secret",
          "provisioning_uri"
        ],
        "title": "TotpEnrollmentRead",
        "type": "object"
      },
      "TotpEnrollmentRequest": {
        "additionalProperties": false,
        "properties": {
          "challenge_token": {
            "title": "Challenge Token",
            "type": "string"
          }
        },
        "required": [
          "challenge_token"
        ],
        "title": "TotpEnrollmentRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "CBPUPSIS API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/admin/me": {
      "get": {
        "description": "Return the MFA-completed Admin identity and query-ready data scope.",
        "operationId": "read_admin_session_api_v1_admin_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminSessionRead"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read Admin Session",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/api/v1/auth/admin/login": {
      "post": {
        "description": "Verify credentials and issue no session until MFA completes.",
        "operationId": "login_admin_api_v1_auth_admin_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminLoginChallenge"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login Admin",
        "tags": [
          "admin-auth"
        ]
      }
    },
    "/api/v1/auth/admin/logout": {
      "post": {
        "description": "Revoke the refresh token and clear the Admin cookie.",
        "operationId": "logout_admin_api_v1_auth_admin_logout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Logout Admin",
        "tags": [
          "admin-auth"
        ]
      }
    },
    "/api/v1/auth/admin/mfa/totp/confirm": {
      "post": {
        "description": "Prove the enrolled seed and issue the first Admin session.",
        "operationId": "confirm_totp_api_v1_auth_admin_mfa_totp_confirm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TotpEnrollmentConfirmRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminTokenPair"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Confirm Totp",
        "tags": [
          "admin-auth"
        ]
      }
    },
    "/api/v1/auth/admin/mfa/totp/enroll": {
      "post": {
        "description": "Return the standard otpauth URI Google Authenticator scans.",
        "operationId": "enroll_totp_api_v1_auth_admin_mfa_totp_enroll_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TotpEnrollmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TotpEnrollmentRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Enroll Totp",
        "tags": [
          "admin-auth"
        ]
      }
    },
    "/api/v1/auth/admin/mfa/verify": {
      "post": {
        "description": "Complete sign-in using Google Authenticator TOTP or a hardware key.",
        "operationId": "verify_mfa_api_v1_auth_admin_mfa_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaVerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminTokenPair"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Verify Mfa",
        "tags": [
          "admin-auth"
        ]
      }
    },
    "/api/v1/auth/admin/refresh": {
      "post": {
        "description": "Rotate an Admin session and renew its hardened cookie.",
        "operationId": "refresh_admin_api_v1_auth_admin_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminTokenPair"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh Admin",
        "tags": [
          "admin-auth"
        ]
      }
    },
...
    "/api/v1/auth/login": {
      "post": {
        "description": "Exchange email and password for an access/refresh token pair.\n\nRate limited against credential stuffing.\n\nThree outcomes, and the difference between them is deliberate:\n\n- **401** \u2014 unknown email, wrong password, or a disabled account, all with\n  one identical message so the endpoint cannot be used to discover which\n  addresses are registered.\n- **403** with ``code: \"email_not_verified\"`` \u2014 the password was correct but\n  the address is unverified. Branch on the code, not the message, and send\n  the user to a resend-verification screen rather than back to login.\n- **200** \u2014 a token pair.",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPair"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "description": "Revoke a refresh token. Idempotent: always succeeds.",
        "operationId": "logout_api_v1_auth_logout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Logout",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "description": "Exchange a refresh token for a new pair, invalidating the old token.",
        "operationId": "refresh_api_v1_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPair"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh",
        "tags": [
          "auth"
        ]
      }
    },
...
  }
}